+++ Assets/Scenes/TestScenes/BrunoCostaTestScene/Scripts/BoxController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5801d05914fd5ad4b89692a8ff9c4e4d
+timeCreated: 1459898192
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
